+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## <small>0.7.3 (2025-06-15)</small>
+
+* refactor: fixed rate limiter for batch requests; ([44ea1b9](https://github.com/EasyLayer/core/commit/44ea1b9))
+
+
+
 ## <small>0.7.2 (2025-06-15)</small>
 
 * refactor: fixed rate limiter and batch requests; ([bbccf98](https://github.com/EasyLayer/core/commit/bbccf98))